moved states into appcontent
